feat: check that root process is in both communicators
To simplify our code we will require for now that the root process
(i.e., rank 0 in MPI_COMM_WORLD) is present in both of the communicators.
One big benefit of this is that we do not have to check whether an
intersection exists between both communicators.

Later on we can think of removing this constraint. If so, we should
check that there is an intersection between the communicators (or perhaps
simply merge them).