feat: add FundamentalType constraint
In the current implementation of shuffle we assume that we
are using one of the fundamental types (e.g., int, float). We make
this now explicit via a new concept; this should help us once we
move to more general implementations that handle user created
datatypes (e.g., structs, classes).