feat: add constraints to Serializable
We need a way to guarantee that we will be able to create an object of
the type shuffle is being called with. This can be a problem for non-aggregate
types, as they might not have a default constructor. So for now, we require that any
non-trivially serializable object implements a create method. In principle this is
only required if the non-aggregate type is also not default constructible.