feat: enable coloring
The coloring enables users to explicitly indicate to which rank a certain
element should go. This can be useful also to represent 2D data in 1D, and do
fancy stuff as scattering blocks of rows/columns among ranks.

In order to enable this feature, several changes were required. Perhaps the largest
include the refactoring of scatter and gather, to remove code duplication.

Finally, we are still making all the work on rank 0.