feat: add initial implementation coloring utilities
We can use coloring to, for example, represent a 1D vector as if it were
2D and then distribute the values using some strategy. This commit
includes a first basic strategy: BlockWise. With it, we can separate our
data in fixed blocks, and assign one block per rank. We can apply it independently
to each dimension; so, for example, having two blocks on each dimension will end up
with a schema that has 4 ranks in total.