fix: make combine row major traversal
We would like to assign ranks by default on a row major basis.
This means that if, for example, a matrix is split in 2X2 blocks, rank 0
should be assigned the elements on a corner, and rank 1 the elements
immediately to the right (i.e., in contrast to column major, on which
case it would get the elements below/above). Since we are using (x, y) coordinates for the blocks,
we need to iterate first over the X to get this effect.