refactor: change attribute function to indicate global scope
The create_coloring function requires both the current global coloring
and, for the 2D case, the global dimensions of the domain. With the
previous names, it was unclear whether those values were indeed
global or local (i.e., relatives to each rank subdomain). Thus,
we add `global_` to address this issue.