feat: make create_coloring return global coloring
While working on a simple example that made use of our library, I noticed that
there was no easy way for the user to get the global coloring required to
create new colorings. Thus, now create_coloring returns this value as well.
The idea is that a user can now create coloring, get the global coloring, and
then use it later on for new shuffles if required.