refactor: encapsulate function to get subdomains
Right now the get_subdomain_dimension and create_coloring assume
a certain order of the subdomains; so, for example, rank 0 always gets
the first subdomain, and rank 1 always the second. Now, what the "first"
and "second" are depends on how we combine the changes on x and y dimensions
(e.g., right now we first increment in X, and then in Y, so that rank 1 will be
to the right on rank 0 and not, for example, to the bottom).

Now, as stated before, we need to keep this order consistent between the functions, so
we encapsulate the logic in a function that is in turn call internally by both of them.