feat: implement shuffle for 2D data structures
While coding the first example application, I noticed that
it was a bit cumbersome for the user to have to re-create a 2D data structure
from the 1D buffer provided by shuffle. It would be a lot cleaner for
shuffle to directly return the 2D variant. This is exactly what is
implemented on this commit.

The main API difference with the 1D version is that, for 2D, we
explicitly require the coloring and the subdomain dimension after shuffle.
The reason for the first is that, unlike in 1D, there is no clear default
behaviour for the function (e.g., should be the default to split row-wise,
column-wise, minimize interfaces, etc.). The reason to require the
subdomain dimensions is to facilitate our lives once we need to return
the result of the shuffle.