build: add `MPIEXEC_PREFLAGS` to test runner
While doing an initial test in docker, I noticed that the tests
were not being build because OpenMPI complained that the code
was about to be run as root. Usually this would be solved by
adding the `--allow-run-as-root` flag, but we had no way to
insert such flags in our build system. Taking a look at the MPI
package documentation, I noticed that the variable
`MPIEXEC_PREFLAGS` could be used for this.

So, after this commit we can happily run this inside a docker
simply by doing:

```
cmake -DMPIEXEC_PREFLAGS=--allow-run-as-root [...]
```

I also updated the documentation to explain this.