build: reduce minimum required version CMake
While trying to run the code on a Rocky Linux 9 instance, I
noticed that only CMake 3.26 was available. Since we are not
using any 3.27 feature, the simplest thing to run the code in
such environment is to simply also accept builds with 3.26.

As a side note, it might be the case that we could reduce this
even more. Nonetheless, for now I think the best approach is
to simply do it on a case by case basis (i.e., if required).