refactor: ignore conan dependencies folder
CLion uses two different short-lived containers for
generating the compilation files and for building. This
mean that by default the conan home directory set up
with all the dependencies on the first step, will not be
available for the second container. As a workaround,
I created the `conan-deps`folder, which is manually
mounted between containers. This folder should nonetheless
not be tracked via GIT.