fix: use MPI_COMM_NULL to determine if rank in comm
When trying to use the library in conjunction with the
dyn_proc demo, I noticed that calls to MPI_COMM_WORLD
would usually cause errors. The reason for this is that
in the Sessions model there is no MPI_COMM_WORLD anymore.
Thus, we redefine in_mpi_comm to simply check whether
the parameter passed was MPI_COMM_NULL. The idea is that
on those demos, when a reduction occurs, the ranks that
are about to leave pass this value as destiny
communicator.

I was a bit surprised that no additional changes were
required to the current demos and/or tests. It is
specially interesting that this seems to work even if
one passes as parameter a communicator to which a certain
rank does not belong; further investigation should be
done to guarantee that we are not introducing here a bug.
If that is the case, the solution would be to make
mandatory the use of MPI_COMM_NULL on cases where the
rank does not belong to a communicator.