feat: add option to enable/disable dyn_proc_demo
The idea is to avoid as much as possible having multiple
branches in the project, as this becomes annoying and
unsustainable with time (e.g., constant need for rebases).
Since the compilation of our dyn_proc_demo requires a
specific version of OpenMPI, this code cannot be compiled
in a normal environment. That was the only thing blocking the
merge of this branch into main. By adding an option, we can
now turn this compilation ON only when we know we have access
to this special OpenMPI version, and leave it without
compilation in all the other cases.