feat: create proper type for Subdomains
Before this commit there was no way for an external
user of the library to determine what kind of range
a subdomain was (e.g., open, closed). The only way was
taking a look at the internal in_range free function. This
of course made things messy. So now instead, we have created
a proper type with a method contains.