doc: indicate how to compile tests on dyn_proc
By default, GTest will try to run the tests using mpirun, which fails on the
cluster. One needs to explicitly change the MPI executable, as well as some
of the flags. Since this was not being indicated on the documentation,
we were getting reports of failures to compile in the dyn_proc
environment.