feat: make Dimension a std::array
After some consideration, I noticed that Dimension was
nothing more than a glorified std::array. Since I still would
like to keep it as a separate type, I opted to make it an alias
of an array of integers. The alternative would be of course
to remove it completely and use instead a std::array<int, N>.