feat: add first implementation shuffle with distribution
Right now every time an user wants to use shuffle in 2D,
they have to separately create a coloring. In order to
simplify this, I would prefer that the user instead directly
indicates the distribution he/she requires. This is a first
step in this direction, as here we implement an initial version
of a 2D shuffle.