refactor: require again coloring attribute
My initial approach to implement coloring-free variants of
shuffle was to have coloring-free versions of scatter. Nonetheless,
it was proven almost necessary to have the "split equally" part
of the code visible in the shuffle function. This means that the functions
inside mpi_utils should always expect a coloring to be indicated.