refactor: use get_global_coloring where possible
All the calls to create_coloring that were using only the global_coloring
could be replaced by a call to get_global_coloring. This is because, for now,
we assume that all data is first gather in root and then scatter elsewhere. This
assumption of course might change after we optimize the information
exchange.