refactor: remove support for serialization
The serialization feature was quite nice, but it was not really being
used anywhere. Since for now I would like to focus on code quality and
improving the efficiency of the library, and the serialization part
adds complexity, I consider best to remove it for now. If at some point
in the future this is required again, then we can think of a proper implementation.