fix: make vertical split work
The main problem was that the code was assuming that
the values of two consecutive ranks will also be consecutive. While
that is true for a horizontal split (i.e., we use row-major memory),
it is not for the general case. So now, we use the old global coloring
to re-create the original matrix in the root before scattering the values.