feat(demo): add funtion to initialize matrix
Before we were initializing the matrix with 0s, which is okay if
we only want to check that the correct number of values is distributed
to each rank. Nonetheless, recent bugs proved that it is useful to
also have unique values on each position, such that we can also
see that the correct values are being distributed.