fix(block-cyclic): make get_num_values return 0 for rank_id larger that num_ranks
Originally I wanted to throw here, but that would break many scenarios. For example,
let's assume we start with a distribution with only one rank and want to move to one with two;
since both ranks will run the checks of whether the correct number of values will be passed,
the code would throw for all but rank 0. A solution would be to check first if the distribution contains the rank, but
this start to add too much boiler plate code.