refactor(shuffle): create separate functions to check conditions
In every variation of shuffle we have different conditions that need to be checked,
and several have to do the same checking (e.g., both in 1D and 2D we need to verify
that the correct number of values is passed). Although they are pretty straight forward,
I was starting to have a lot of boilerplate code, making the functions even more complex. Thus,
I decided to abstract the logic in separate functions that should throw if conditions are not
met.