refactor(shuffle): give more information exception message
While catching a bug I noticed that it would be nice for the `check_correct_num_values_provided`
to inform the expected number of values and what it actually got. Since now the message will depend on
the scenario, I modified also the tests to check for the substring instead of doing a full match.