fix(utils): solve issue calculating number of values multiple dimensions
I observed that for a 2x2 domain `split num_values_in_rank` was returning 4 for ranks
2 and 3. The issue here was that, since the distribution is only aware of the number of
ranks in their "direction" (i.e., the distribution in X is only aware that there are N ranks
in X, and has no idea how many are in Y), asking for the number of values of rank, let's say 3 in a 2x2,
would return 0, which in turn would be transformed to the total number of values per our
previous algorithm.

Now, instead, we simply get the number of values in the respective row and column.