refactor(mpi-utils): move core functions to separate definition file
I am in the process of refactoring mpi utils. In particular, there are a set
of core functions that I am sure will remain as part of the header. Of those, the ones that
are not templated, have been moved to a definition file. The main reason for this is to
reuse their functionality.