fix(utils): fix issue with reorder_values semantics
While doing the refactoring I thought that reorder_values was
simply meant to return the values indicated by the indices vector (e.g.,
[2, 1, 3] would mean give the second, first, and third value), but instead
the indices vector indicated the new order (e.g., in the previous example,
send the first element to the second position, the second to the first, and
keep the last one in the last position). I fixed that issue.