opt: introduce first version of point to point communication scheme
Until this point, a shuffle required to send all data to root, from where the values
were scattered to the respective ranks. This commit introduces the first version of a
point to point scheme that avoids this duplication.

Initial GBench results show an improvement with respect to the previous scheme.
Nonetheless, some refactoring is still required.