fix(benchmark): passing nullptr to pdgeadd even if rank has data
As it was written, the source data passed to `pdgeadd_` was set to nullptr to
every rank but rank 0, regardless of whether that specific rank had data or
not. This was okay before, where we were only testing redistributions from one
to many, but it is problematic if we want to redistribute from many to
many.