opt: return immediately if the two distributions are the same
Checking the benchmarks of ScaLAPACK and reshuffle, I noticed that the time
of the former was basically nothing. The benchmarks right now are doing something
a bit stupid: going changing from one distribution to the same. So, ScaLAPACK probably
checks for this and makes no distribution if this is the case. To have comparable results,
now I also check for this.