fix: wrong calculation number of values last block
We were calculating the number of values of the last block, which is
assigned to the last rank, as the remainder (i.e., modulo) of the
division of the number of values and the number of blocks. The problem
is that this does not work when the number of values is perfectly
divisible, as in such cases such reminder is 0, which of course is not
the size of the block.