opt: use blocks instead of coloring
This is a big commit, where I introduce a variant of reshuffle (for now
just in the `dev` namespace). This variant uses the blocks, instead
of the coloring, to determine for each rank what data to send and to
receive.

The previous approach has huge benefits in terms of performance, and
also in terms of code quality, over our previous approach. Before, we
had to calculate for each element the source and destiny. In practice,
that meant that the time complexity of the coloring approach did not
scale well.

There is still a long road to have a complete implementation. For
example, the current prototype only works for 1D buffers. Also, as it
is, the same communication must be used by both the initial and final
context. We will also have to update our documentation and change our
examples once we finish the transition.