feat: enable partial support for different communicators in contexts
Before this commit, I was simply using the communicator from the initial
context as a base communicator. This, of course, was silly, but
considering that all the tests were using MPI_COMM_WORLD for both
contexts, it was good enough.

Now, the two contexts can have different communicators, as long as one
is a subcommunicator of the other.

Working with partially disjoint communicators will be pushed for the
future, as it might require changing parts of the algorithm and coming
with workarounds for MPI.