refactor: separate mpi core utilities from communication utilities
Continuing with my effort of clearing defining what the core utilities
of mpi are, I now separated functions related to communication and
functions related to more general tasks into two headers. As a rule of
thumb, if a routine is used to exchange information, then it is not a
core utility.