feat: make shuffle work with subcomms that are not aligned
Until now, we have two restrictions when shuffling from contexts with
different communicators:

1. Either the initial or final communicator should be a subcommunicator
of the other.
2. If a rank is in both communicators, then its rank id has to be the
same.

The above meant that if we assume a global communicator with ranks
{0, 1, 2, 3, 4}, then:

- {0, 1, 2, 3} and {0, 1} are valid and could be used.
- {0, 1, 2, 3} and {1, 2} are not valid, as rank 0 in the second
communicator would be rank 1 in the first one.

With this commit we generalize a bit more and remove the second
condition altogether. We do so by adding an Intercommunicator class,
whose job is to:

 1) Create a common communicator that can be used by ranks
from the initial and final context
 2) Act as translator between the two. This allows getting the rank in
 the common communicator from the rank in the initial and final
 communicators, and the other way around.