feat: templatize ProcessorGrid for arbitrary dimensions
This is the first commit on my quest to enable reshuffling of data in
arbitrary dimensions (e.g., 2D matrices). As a first step, I transformed
ProcessorGrid to store a Dimension object instead of only the number
of processors.

To keep things simple, I have for now set ProcessorGrid<1> in all the
classes that use ProcessorGrid (e.g., BlockCyclic). This, of course,
will be updated as I go up the chain changing more and more classes.