feat: replace cartesian view with functions
I tried several times to use the cartesian grid view, but it was always
problematic. The pipeline was not supported (this also applies to
"proper" implementations, such as the one in v3-ranges), it would
produce tuples that needed to be unpacked, and, most importantly, the
code was hard to understand. Thus, I opted to implement the cartesian
product for up to 3 dimensions and assume the cost of copying those
values.