fix: mapping coordinates
In a "normal" linear algebra fashion, when we have a 2x1 matrix, we
have a matrix with two rows and one column (i.e., horizontal partition).
But in our project, when we create a Dimensions object (which is in turn
used by Coordinates to calculate indices and coordinates), we indicate
the number of partitions in the respective dimension. So, for example,
a Dimensions{2, 1} indicates two partitions in x and 1 in y; so, we have
a domain with two columns and one row (i.e., vertical partition). This
is, of course, the opposite of the classic interpretation.

Our algorithms were assuming the Linear Algebra interpretation, not the
actual intended meaning with our Dimensions. With this commit we fix
that. As it can be seen in the code, we simply needed to invert the
Dimensions object.