feat: templatize GridLayout for multiple dimensions
In this commit, I have modified GridLayout (and the code it depends on
it) to work with multiple dimensions. The class itself can work with
arbitrary dimensions, but we are currently limited by 3 by the
cartesian_product functions.

For now, all functions that consume/use GridLayout up the chain are
still not templetazied, and thus assume a GridLayout of only 1
dimension.