feat: templatize BlockCyclic for arbitrary dimensions
For now, the only limiting factor is that GridLayout can only be
instantiated with up to 3 dimensions (i.e., it is limited in turn by
the fact that we can only do cartesian products in up to 3D).

As before, any class/function that uses BlockCyclic is for now assuming
a 1D domain. This will change as we move up in the hierarchy.