feat: make GridOverlay a class instead of a struct
I might need to calculate the coordinates of the processors in the
target grid, so I thought about simply doing it in GridOverlay. That
requires nonetheless making it a class instead of a struct (i.e., I
want to avoid requiring the processor coordinates in the constructor as
a parameter).