fix: avoid throwing when ProcessGrid has fewer ranks than MPI_Comm
It might happen that the ProcessorGrid has fewer ranks than the
MPI_Comm used in a specific context.

For example, let's assume you are running an application with 2 ranks
but want to split only a 1D buffer that currently is stored only by
rank 0. In this example we would go from a BlockCyclic distribution
with one rank in its processor grid, to one with two.

So far, so good. But now, let's assume that we want to use
MPI_COMM_WORLD in both the initial and final context. This means that,
with our current algorithm, at some point we will try to get the
processor coordinates of rank 1 for the initial context (i.e., the
second rank) which in turn was created with a ProcessorGrid of 1.

Before we moved our code to coordinates, we were simply returning an
INVALID_RANK_ID, and that was okay. But now, we need to call the
map_index function, which would throw when requested to get the
coordinates of rank 1. Now, instead, we make map_index and map_indices
return an optional. The callee can then decide what to do if the
optional is empty. In our case, the callee is ProcessorGrid, and for
those cases will return some invalid coordinates (or INVALID_RANK_ID,
for get_processor_id), mimicking what was being done before.

As a side note, I observed some weird behavior on some tests, where
the code would not simply crash, but the shuffle function would be
called again. I am not sure what was triggering this behavior.