fix: wrong unpack of owner coordinates in get_local_grid()
It is tricky right now how to unpack the coordinates of the owner of a
MultidimensionalBlock. As explained in get_owner_coordinates, we need
to extract the owners of each interval in the opposite order in which
they are stored. It turns out we were not doing that when returning the
local grid of a processor.

The easiest solution is to simply use the already existing
get_owner_coordinates.