fix: cartesian product should change first x, then y, then z
This is easier to explain in 2D. Let's assume we have an array of
vectors, where the first vector corresponds to the Blocks in X and
the second to the Blocks in Y. When we combine them using the
cartesian product, we want to generate a vector of
MultidimensionalBlocks (MDs) (i.e., vector of arrays), such that they
are ordered in the same way as the processor grid: using a row major
scheme. This in turn means that we need to move first on the X-axis
(i.e., the horizontal one), and after in the Y-axis (i.e., vertical).
Thus, if we produce MDs of the form (b1, b2), b1 should be iterated
first.