feat: use standard lineal algebra notation for dimensions
As stated in previous commits, before our code was assuming an inverted
representation of the dimension of a matrix. For example, Dimension{2,
1} was equivalent to a Matrix 1x2. The reasoning for this was that at
the beginning it was easiest for me to think in terms of how many splits
I wanted per dimension starting with x, not with y.

Sadly, as I move forward with the library, I am starting to notice that
this can cause confusion and can complicate the code. For example, I had
to modify the standard cartesian product to iterate first over the first
element of the tuple when the normal thing is to start by the right-most
element. I also had to invert the dimensions when calculating the
coordinates of a certain rank. Now the latest thing I found is that
std::mdspan, which I am considering to utilize, requires in its
constructor the dimensions in the classic linear algebra order. So,
enough is enough, and I decided to switch to the standard notation.

From now on, for example, a processor grid with Dimension{2, 1} will
have to rows and one column.