refactor: rename rank variables in shuffle
As usual, after being some weeks without working on the code base and
returning, I found some naming a bit strange. In particular, I was
confused about `initial_rank` and `final_rank`, and they usage by the
`get_send_and_receive_blocks`. The naming suggested that I was only
packing values owned by the range of ranks starting at initial and
finalizing at final. That was obviously not the case, and instead I
meant by those names the rank id on the initial grid and the rank id
on the final grid. Thus, to make life of future Santiago easier, I
just renamed the variables to clarify this.