refactor: simplify last step `get_send_and_receive_blocks`
The last step corresponds to removing duplicated ranges (which may
appear if a rank owns multiple blocks that share the same extension in
one dimension) and joining the values to have the local indices of the
data. Now that we have a separate function for removing duplicates,
we can easily merge both operations into the same transform.