feat(build): disable compilation of Costa and ScaLAPACK by default
Now that other people might start to use the library, I think it is best
to leave the requirements to the bear minimum, so that they can easily
compile our code. Since an external user does not really care about
Costa or ScaLAPACK (which are only used to run some internal
benchmarks), I took the decision to disable their use by default.