fix(build): make sure gtest builds gmock
While trying to compile the code with GCC on my Mac, I observed that
gmock was not being compiled (i.e., I got errors of the type "Gmock
not found"). Although the conan recipe for gtest sets the build_mock
as True by default, I opted to set its value in our `conanfile.txt`.
That seemed to fix that particular issue.