feat: add C++ concepts required to support serialization
We are on the road of (once more) supporting serialization of
user-defined datatypes. As a first step, I have created some useful
concepts to determine when a datatype needs serialization (i.e.,
serialization has an overhead, so we do not want to do this to, for
example, a vector of integers). If a type requires serialization,
then we can also check if it is supported.

As before (see commit 8ba35cc8c30b64140ec85d4c5dfbb852d77ac8de), we
will use zpp-bits for this.