feat: implement exchange_values for user defined types
At this point I am unsure whether the best choice is to specialize
exchange_values or the functions it is using (e.g., get_send_package)
to work with custom datatypes. It seems that the easiest option for
now is to modify the logic of exchange_values, so I am doing that.

This is still in an early stage, and there are a couple of TODOs
to finish and code to clean, but at least the first test is passing.