refactor: differentiate send and receive communication packages
At first, I thought that a single structure would suffice to represent
both sending and receiving packages. After all, the only purpose was
to have a buffer and a description of the buffer. Nonetheless,
there are important differences that are starting to be evident in terms
of how (1) the struct is built and (2) how it is serialized. Thus,
I think best of having two distinct structs.