refactor: use maps instead of vector of blocks
There were several parts of the data exchange routines where it was
unclear whether blocks were grouped by processor or not. At first,
I attempted to solve this by naming variables with `_grouped_by_` when
that was the case, but this was certainly cumbersome.

Nonetheless, the thing that finally made me refactor this was a line
in our exchange_values function:

```
const auto &received_block = grouped_blocks_to_receive[source];
```

That was super error-prone, as it basically assumed that we were
always receiving from consecutive ranks starting at 0. For example,
if rank 0 were only to receive from rank 1, the code above would break.

It is true that the case above would be impossible right now, as
reshuffle only supports BlockCyclic distributions; but it is also true
that we should support more distributions in the near future. Thus,
to save myself a headache, I just went ahead and implemented the change.

In my opinion, now the code is clearer as before, and we also were
able to avoid many indirections of the form block.get_interval().blah...