fix(build): indicate that we need zpp_bits for code linking our library
For the moment, code that wants to link to our library needs to have
installed zpp_bits. The reason for this is that we use zpp_bits to
serialize user-defined types, which requires the modification of the
user code (i.e., add explicit serialization).

We had sadly forgotten to modify our Config.cmake.in to express this,
which caused errors when trying to link our code with AutoPas. This
should fix that.

Additionally, I made a note in our documentation.