feat: simplify the exchange of data via simpler send/recv functions
The code to exchange data had diverged depending on whether the data
needed to be serialized or not. The root of the problem was that
we were explicitly calculating the size of the receiving vector, and
this calculation changed when serialization was required.

Now, instead of calculating the receiving buffer beforehand, we use
MPI_Probe() and MPI_Get_count() for this. We have also added several
mpi helper functions to send and receive data.

Thanks to all of this, now the logic of exchange_values is the same
regardless of whether serialization is required or not, greatly
simplifying the code.