refactor: move overload `get_num_elements` to multidimensional_block.hpp
We had a second version of `get_num_elements` that took a vector of
MultidimensionalBlocks defined in multidimensional_data.hpp. I think
best to have all possible overloads in the same file, and thus decided
to move it to multidimensional_block.hpp.

I have also added a missing test for this functionality.