refactor(benchmark): rename reshuffle benchmark
Back when we had two shuffle versions (one using the old coloring scheme
and the other with the block algorithm), we had a normal reshuffle
benchmark and a new one. After the coloring scheme was deprecated,
we ended up only with the new reshuffle. After this long, it feels
weird to have a `new_reshuffle_benchmark.cpp`, so I finally renamed it
to simply `reshuffle_benchmark.cpp`.