fix: avoid segfault when non-root passes non-empty map
In the block_scatter version of fundamental types we had no check of
whether a specific rank was the root rank when dealing with the values
per rank map. This is okay, as the values vector is anyway not
accessed.

Now, the block_scatter version for aggregate datatypes needs to
serialize the values. Before, we were not only doing this for the root
rank, which of course lead to segfaults.