fix(test): interchange vertical and horizontal partition tests
In commit 34ab5c1f812fbdf0e6237ab15c35e8fa571fef69 we change the
representation of our partitions to follow the more standard linear
algebra convention that `2x1` -> `{2, 1}` represents a system with two
rows and one column. The issue is that the tests in grid_layout_test
and shuffle_test were not updated accordingly.

For example, we had` CanReturnTheOwnerOfABlockIn2DHorizontalPartition`
which despite having horizontal partition on the name, was creating a
processor grid with `{2, 1}`, which as explained, would represent
instead a vertical partition.