fix(grid_layout): make the order of blocks in the constructor consistent
As explained in the previous commit, we have the convention that any
N-dimensional representation has the most "significant bit" on the left,
and the most insignificant on the right. So, for example, a coordinate
(1, 0) in a 2x2 grid corresponds to the third element located in the
bottom left, and NOT the second in the top right.

Now, a multidimensional block takes as an argument an array of blocks.
In a 2D scenario one should pass to the constructor {y-block, x-block},
as shown in multidimensional_block_test.cpp.

The problem was that GridLayout was inverting this convention. It
expected instead to get, for 2D for example, {x-block, y-block}. This,
of course, was quite confusing and error-prone. For example, we had to
invert the order of the coordinates when calling
get_processor_coordinates.

Now I changed the code such that the convention is kept also for
GridLayout.