opt: use a simple scatter when going from one rank to block-wise
Going from a distribution assigned to one rank to a block-wise
distribution is equivalent to simply doing a scatter. This saves us
some considerable amount of time, as there is no need to copy/organize
the values again once they have arrived.

We also add a new benchmark for scattering values. We can still optimize
this further in 1D by simply passing the buffer to the scatter function.