feat(scripts): update plotting scripts
This constitutes almost a complete rewrite of our original plotting
scripts. The original version was developed in a bit of a rush, as I
simply wanted to check some benchmark results. With time, it became
more and more difficult to extend them. In particular, I struggled to:

1. Make the compare script be able to compare more than two files.
2. Change titles depending on grouping.

Thus, I decided it was time for a refactor.