feat: implement a simpler way to get a Grid Overlay
The current way to get an overlay is quite cumbersome and error-prone.
For example, people need to know that the multiblocks stored in the
overlay have as owners the ids of the origin.

With this commit I introduce a new definition for GridOverlay (called
GridOverlayDev for now, as it has yet to replace the old one). The code
is indeed simpler and easier to understand than the previous class.