fix: make_contiguous algorithm MultidimensionalBlocks
If we have blocks that are already aligned in one dimension, we expect
the algorithm to keep it that way. What was happening before instead,
was that the algorithm was adding them up. So, if we had two blocks with
position in x {0, 2}, the first one was keeping {0, 2}, but the second
one was getting {2, 4}, which is wrong.

This commit addresses this particular issue.