fix(tests): vertical and horizontal shuffle tests were swap
The tests named "horizontal" were actually using vertical partitions,
and the "vertical" ones, horizontal partitions. This is easy to verify
if we consider that the `ProcessorGrid<2>{{1, 2}`, which generates a
processor grid
with 1 row and two columns (i.e., vertical partition), was used for the
horizontal test.

I fixed this by simply exchanging the names.