feat: add `is_block_wise` to DataDistribution interface
We had a function `is_block_wise_distribution` that would take a
BlockCyclic distribution and return true if the data was split in a
block-wise manner. This is an important optimization, and thus we should
check this also for future types of distributions. Thus, I decided to
add it to the interface.