refactor: move several classes outside internal namespace
The DataDistribution interface forces us to implement `get_grid_layout`,
which returns a `GridLayout`. The problem we had is that, although
`DataDistribution` belongs to the main `reshuffle` namespace,
`GridLayout` belonged to the internal one. This is of course not so
nice.

As a solution, I decided to move `GridLayout` out of `internal`. This in
turn affected other classes, that also had to be moved:

- `Block`
- `Coordinates`
- `MultidimensionalBlock`
- 'LeftClosedRange'

`Coordinates` and `MultidimensionalBlock` have functions (not methods)
that operate on them. These are only meant to be used internally. Thus,
thus functions remain in the `reshuffle::internal`.