feat: add Interval as an alias for LeftClosedRange
I consider `Interval` to be a better name in general that
LeftClosedRange. I originally used the second one just to be explicit
about what was included in the interval and what not. The alias brings
the benefit of simplifying code, while still making sure its easy to
check what an interval is.